+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules.hs
@@ -2,7 +2,7 @@ module Cardano.Ledger.Conway.Rules (
   module Cardano.Ledger.Conway.Rules.Cert,
   module Cardano.Ledger.Conway.Rules.Deleg,
   module Cardano.Ledger.Conway.Rules.Pool,
-  module Cardano.Ledger.Conway.Rules.VDel,
+  module Cardano.Ledger.Conway.Rules.GovCert,
   module Cardano.Ledger.Conway.Rules.Certs,
   module Cardano.Ledger.Conway.Rules.Enact,
   module Cardano.Ledger.Conway.Rules.Epoch,
@@ -10,7 +10,7 @@ module Cardano.Ledger.Conway.Rules (
   module Cardano.Ledger.Conway.Rules.NewEpoch,
   module Cardano.Ledger.Conway.Rules.Tickf,
   module Cardano.Ledger.Conway.Rules.Ratify,
-  module Cardano.Ledger.Conway.Rules.Tally,
+  module Cardano.Ledger.Conway.Rules.Gov,
   module Cardano.Ledger.Conway.Rules.Utxos,
   module Cardano.Ledger.Conway.Rules.Utxow,
 )
@@ -21,12 +21,12 @@ import Cardano.Ledger.Conway.Rules.Certs
 import Cardano.Ledger.Conway.Rules.Deleg
 import Cardano.Ledger.Conway.Rules.Enact
 import Cardano.Ledger.Conway.Rules.Epoch
+import Cardano.Ledger.Conway.Rules.Gov
+import Cardano.Ledger.Conway.Rules.GovCert
 import Cardano.Ledger.Conway.Rules.Ledger
 import Cardano.Ledger.Conway.Rules.NewEpoch
 import Cardano.Ledger.Conway.Rules.Pool
 import Cardano.Ledger.Conway.Rules.Ratify
-import Cardano.Ledger.Conway.Rules.Tally
 import Cardano.Ledger.Conway.Rules.Tickf
 import Cardano.Ledger.Conway.Rules.Utxos
 import Cardano.Ledger.Conway.Rules.Utxow
-import Cardano.Ledger.Conway.Rules.VDel